+++ routes/my-team.routes.js
@@ -4,6 +4,7 @@ const getMySupervisorsController = require('../controllers/account-setting/my-te
 const getMyAgentsController = require('../controllers/account-setting/my-team/get-my-agents.controller');
 const removeAccessFromMyTeam = require('../controllers/account-setting/my-team/remove-access-from-myteam.controller');
 const fetchAccessRequests = require('../controllers/account-setting/my-team/fetch-acces-requests.controller');
+const changeAccessRequestStatus = require('../controllers/account-setting/my-team/change-access-request-status.controller');
 
 
 const myTeamRouter = Router();
@@ -13,5 +14,6 @@ myTeamRouter.get('/get-my-supervisors/:user', getMySupervisorsController);
 myTeamRouter.get('/fetch-access-request/:user', fetchAccessRequests);
 myTeamRouter.get('/get-my-agents/:user', getMyAgentsController);
 myTeamRouter.put('/remove-access-request/:user', removeAccessFromMyTeam);
+myTeamRouter.put('/change-access-request-status/:user/:id', changeAccessRequestStatus);
 
 module.exports = myTeamRouter;
